refactor(CoherentCacheTests): Split cache tests in files to keep a clean structure
