+++ src/front/macOS/kDriveTests/Cache/CoherentCacheTests/CoherentCacheUserTests.swift
@@ -1,0 +1,85 @@
+/*
+ Infomaniak kDrive - Desktop
+ Copyright (C) 2023-2025 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Foundation
+@testable import kDriveCore
+import OrderedCollections
+import Testing
+
+struct CoherentCacheUserTests {
+    @Test(.timeLimit(.minutes(1)))
+    func getUserInCache() async throws {
+        // GIVEN
+        let user = CacheData.expectedUser
+        let cache = ServerCoherentCache()
+        #expect(await cache.getUser(apiId: CacheData.expectedUserAPIId) == nil)
+
+        // WHEN
+        await cache.addUser(user)
+
+        // THEN
+        #expect(await cache.getUser(apiId: CacheData.expectedUserAPIId) == user)
+        #expect(await cache.getUser(dbId: CacheData.expectedUserDbId) == user)
+    }
+
+    @Test(.timeLimit(.minutes(1)))
+    func removeUserInCacheFromDbId() async throws {
+        // GIVEN
+        let user = CacheData.expectedUser
+        let cache = ServerCoherentCache()
+        #expect(await cache.getUser(dbId: CacheData.expectedUserDbId) == nil)
+        await cache.addUser(user)
+        #expect(await cache.getUser(dbId: CacheData.expectedUserDbId) == user)
+
+        // WHEN
+        await cache.removeUser(dbId: CacheData.expectedUserDbId)
+
+        // THEN
+        #expect(await cache.getUser(dbId: CacheData.expectedUserDbId) == nil)
+    }
+
+    @Test(.timeLimit(.minutes(1)))
+    func updateUserInCache() async throws {
+        // GIVEN
+        let user = CacheData.expectedUser
+        let cache = ServerCoherentCache()
+        #expect(await cache.getUser(dbId: CacheData.expectedUserDbId) == nil)
+        await cache.addUser(user)
+        #expect(await cache.getUser(dbId: CacheData.expectedUserDbId) == user)
+
+        // WHEN
+        await cache.updateUser(CacheData.updatedUser)
+
+        // THEN
+        guard let userByDbId = await cache.getUser(dbId: CacheData.expectedUserDbId) else {
+            Issue.record("We should be able to fetch a user from db id")
+            return
+        }
+
+        #expect(userByDbId != CacheData.expectedUser)
+        #expect(userByDbId == CacheData.updatedUser)
+
+        #expect(userByDbId.userId == CacheData.updatedUserAPIId, "The API id should change")
+        #expect(userByDbId.name == CacheData.updatedUserName, "The user name should change")
+
+        #expect(await cache.getUser(apiId: CacheData.expectedUserAPIId) == nil,
+                "Should not be able to fetch an object with the old API id")
+        #expect(await cache.getUser(apiId: CacheData.updatedUserAPIId) == CacheData.updatedUser,
+                "Should be able to fetch an object with the old API id")
+    }
+}
